+++ lib/splunk-sdk-ruby/xml_shim.rb
@@ -113,5 +113,16 @@ module Splunk
       end
     end
   end
+
+  ##
+  # Escape reserved XML entities in a string.
+  #
+  # To match the behavior of splunkd, we only escape &, <, and >, not single
+  # and double quotes. This means we have to write the behavior ourselves,
+  # since both REXML and Nokogiri also escape both kinds of quotes.
+  #
+  def self.escape_string(str)
+    str.gsub(/&/, "&amp;").gsub(/</, "&lt;").gsub(/>/, "&gt;")
+  end
 end