+++ test/test_resultsreader.rb
@@ -19,21 +19,23 @@ class TestResultsReader < Test::Unit::TestCase
     n_results = 0
     reader.each_with_index do |result, index|
       n_results += 1
+      expected_fields = expected["results"][index]["fields"]
       # The assert of the full data structure below works, but
       # by default Test::Unit doesn't print the diff of large
       # data structures, so for debugging purposes it's much
       # nicer to have each key checked individually as well.
-      expected["results"][index].each_entry do |key, value|
+      expected_fields.each_entry do |key, value|
         assert_equal([index, key, value],
                      [index, key, result[key]])
       end
-      assert_equal(expected["results"][index], result)
+      assert_equal(expected_fields, result)
+      assert_equal(expected["results"][index]["RAW_XML"], result.segmented_raw())
     end
     assert_equal(expected["results"].length, n_results)
   end
 
-  test_data = JSON::parse(open("test/resultsreader_test_data.json").read())
-  export_data = JSON::parse(open("test/export_test_data.json").read())
+  test_data = JSON::parse(open("test/data/resultsreader_test_data.json").read())
+  export_data = JSON::parse(open("test/data/export_test_data.json").read())
 
   xml_libraries.each do |xml_library|
     test_data.each_entry do |version, tests|
@@ -53,7 +55,8 @@ class TestResultsReader < Test::Unit::TestCase
       test_name = "test_#{xml_library}_#{version.gsub(/\./, "_")}_sans_preview"
       define_method(test_name.intern) do
         Splunk::require_xml_library(xml_library)
-        file = File.open("test/data/export/#{version}/export_results.xml")
+        raw_file = File.open("test/data/export/#{version}/export_results.xml")
+        file = Splunk::ExportStream.new(raw_file)
         reader = MultiResultsReader.new(file)
         found = reader.final_results()
         expected = tests["without_preview"]
@@ -64,7 +67,8 @@ class TestResultsReader < Test::Unit::TestCase
       test_name = "test_#{xml_library}_#{version.gsub(/\./, "_")}_with_preview"
       define_method(test_name.intern) do
         Splunk::require_xml_library(xml_library)
-        file = File.open("test/data/export/#{version}/export_results.xml")
+        raw_file = File.open("test/data/export/#{version}/export_results.xml")
+        file = Splunk::ExportStream.new(raw_file)
         multireader = MultiResultsReader.new(file)
         n_results_sets = 0
         readers = []
@@ -91,7 +95,7 @@ class TestResultsReader < Test::Unit::TestCase
           readers = []
           multireader.each_with_index do |rr, index|
             readers << rr
-            expected = tests["nonreporting"][index]
+            expected = tests["nonreporting"]
             assert_results_reader_equals(expected, rr)
             n_results_sets += 1
           end